[Backend.CSharp] factor static Compile() method into caller
And rename class to Constants now that no business logic resides in it.

Additionally, remove global state mutation on those fields.
